Move private git utilities into git_utilities
